Blend Strafe Animations, Crosshair Spread, and Pickup Widger
Added new animations for strafing and the ability to blend between each direction based on Player and crosshair direction. Jump animation. The ability for the Crosshair to spread based on firing, jumping, and moving. As well as, the beginning of the weapon Pickup Widget UI for picking up guns in the game